+++ src/app/clients/clients-view/clients-view.component.ts
@@ -108,7 +108,7 @@ export class ClientsViewComponent implements OnInit {
    * Performs action button/option action.
    * @param {string} name action name.
    */
-  doAction(name: string) {
+  doAction(name: string, countryId: string | null) {
     switch (name) {
       case 'Assign Staff':
       case 'Close':
@@ -126,7 +126,7 @@ export class ClientsViewComponent implements OnInit {
       case 'Add Charge':
       case 'Create Self Service User':
       case 'Client Screen Reports':
-        this.router.navigate([`actions/${name}`], { relativeTo: this.route });
+        this.router.navigate([`actions/${name}/${countryId}`], { relativeTo: this.route });
         break;
       case 'Unassign Staff':
         this.unassignStaff();